# docs/openapi(2) (1).yaml
openapi: 3.0.3
info:
  title: Onidel Cloud API - OpenAPI 3.0
  description: |-
    Welcome to the **Onidel Cloud API** documentation. This API allows developers to seamlessly integrate with the Onidel Cloud platform, enabling them to manage cloud resources efficiently and securely. With our API, you can automate processes, retrieve detailed account information, and manage virtual machines (VPS), networking, storage, and more.
  termsOfService: https://docs.onidel.com/policies/terms-of-service
  contact:
    email: support@onidel.com
  version: 1.2.0
servers:
  - url: https://api.cloud.onidel.com
...
paths:
...
  /network/vpcs:
    get:
      tags:
        - VPC
      summary: List VPCs
      description: Retrieve all VPCs for a team
      operationId: getVPCs
      parameters:
        - name: team_id
          in: query
          description: Team ID to filter VPCs. Default team will be used if not provided.
          required: false
          schema:
            type: string
            format: uuid
        - name: location
          in: query
          description: Filter VPCs by location
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VPCs'
        '401':
          description: Unauthorized
      security:
        - api_key: []
    post:
      tags:
        - VPC
      summary: Create a new VPC
      description: Create a new Virtual Private Cloud network
      operationId: createVPC
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewVPC'
      responses:
        '201':
          description: VPC created successfully
        '400':
          description: Bad request - missing or invalid parameters
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - minimum VM required or max VPC limit reached
        '404':
          description: Location not found
        '503':
          description: Service unavailable - resource not available
      security:
        - api_key: []
  /network/vpcs/{vpc_id}:
    get:
      tags:
        - VPC
      summary: Get VPC by ID
      description: Retrieve a specific VPC by its UUID
      operationId: getVPC
      parameters:
        - name: vpc_id
          in: path
          description: VPC UUID
          required: true
          schema:
            type: string
            format: uuid
        - name: team_id
          in: query
          description: Team ID. Default team will be used if not provided.
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VPCDetail'
        '401':
          description: Unauthorized
        '404':
          description: VPC not found
      security:
        - api_key: []
    patch:
      tags:
        - VPC
      summary: Update VPC
      description: Update an existing VPC's name, description, or shared teams
      operationId: updateVPC
      parameters:
        - name: vpc_id
          in: path
          description: VPC UUID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVPC'
      responses:
        '200':
          description: VPC updated successfully
        '400':
          description: Bad request - invalid name, description, or shared teams
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - shared VPC is attached to VMs
        '404':
          description: VPC not found
      security:
        - api_key: []
    delete:
      tags:
        - VPC
      summary: Delete VPC
      description: Delete a VPC (must not have any attached VMs)
      operationId: deleteVPC
      parameters:
        - name: vpc_id
          in: path
          description: VPC UUID
          required: true
          schema:
            type: string
            format: uuid
        - name: team_id
          in: query
          description: Team ID. Default team will be used if not provided.
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: VPC deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - VPC has attached VMs
        '404':
          description: VPC not found
      security:
        - api_key: []
...
  /network/firewalls/{firewall_id}/rules:
    get:
      tags:
        - Firewall Rule
      summary: List Firewall Rules
      description: Retrieve all rules in a firewall group
      operationId: getFirewallRules
      parameters:
        - name: firewall_id
          in: path
          description: Firewall Group UUID
          required: true
          schema:
            type: string
            format: uuid
        - name: team_id
          in: query
          description: Team ID
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FirewallRules'
        '401':
          description: Unauthorized
        '404':
          description: Firewall group not found
      security:
        - api_key: []
    post:
      tags:
        - Firewall Rule
      summary: Create Firewall Rule
      description: |-
        Add a new rule to a firewall group.

        Special subnet values: `Cloudflare`, `Onidel`, `HetrixTools`, `CloudFront`, `UptimeRobot`.
        To reference an IP list, use `list:{ip_list_id}` as the subnet value.
      operationId: createFirewallRule
      parameters:
        - name: firewall_id
          in: path
          description: Firewall Group UUID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewFirewallRule'
      responses:
        '201':
          description: Firewall rule created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FirewallRuleResponse'
        '400':
          description: Bad request - invalid parameters or duplicate rule
        '401':
          description: Unauthorized
        '404':
          description: Firewall group not found
      security:
        - api_key: []
  /network/firewalls/{firewall_id}/rules/{rule_id}:
    get:
      tags:
        - Firewall Rule
      summary: Get Firewall Rule by ID
      description: Retrieve a specific firewall rule
      operationId: getFirewallRule
      parameters:
        - name: firewall_id
          in: path
          description: Firewall Group UUID
          required: true
          schema:
            type: string
            format: uuid
        - name: rule_id
          in: path
          description: Firewall Rule UUID
          required: true
          schema:
            type: string
            format: uuid
        - name: team_id
          in: query
          description: Team ID
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FirewallRuleResponse'
        '401':
          description: Unauthorized
        '404':
          description: Rule not found
      security:
        - api_key: []
    patch:
      tags:
        - Firewall Rule
      summary: Update Firewall Rule
      description: Update a firewall rule's description (max 255 characters)
      operationId: updateFirewallRule
      parameters:
        - name: firewall_id
          in: path
          description: Firewall Group UUID
          required: true
          schema:
            type: string
            format: uuid
        - name: rule_id
          in: path
          description: Firewall Rule UUID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                team_id:
                  type: string
                  format: uuid
                desc:
                  type: string
                  maxLength: 255
                  example: "Allow HTTPS traffic"
              required:
                - desc
      responses:
        '200':
          description: Rule updated successfully
        '400':
          description: Bad request - description too long
        '401':
          description: Unauthorized
        '404':
          description: Rule not found
      security:
        - api_key: []
    delete:
      tags:
        - Firewall Rule
      summary: Delete Firewall Rule
      description: Delete a firewall rule from a group
      operationId: deleteFirewallRule
      parameters:
        - name: firewall_id
          in: path
          description: Firewall Group UUID
          required: true
          schema:
            type: string
            format: uuid
        - name: rule_id
          in: path
          description: Firewall Rule UUID
          required: true
          schema:
            type: string
            format: uuid
        - name: team_id
          in: query
          description: Team ID
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Rule deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Rule not found
      security:
        - api_key: []
...
components:
  schemas:
...
    VPCTeam:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: c1d45377-b2a9-4407-ab8d-6909c34dfaac
        name:
          type: string
          example: "IT Development"
    VPC:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: f5a3e2d8-9876-4abc-1234-543210fedcba
        name:
          type: string
          example: "Production VPC"
        description:
          type: string
          example: "VPC for production environment"
        date_created:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        status:
          type: string
          enum:
            - active
            - building
            - deleting
          example: "active"
        location:
          type: string
          example: "sydney"
        v4_subnet:
          type: string
          example: "10.0.0.0"
        v4_subnet_mask:
          type: string
          example: "24"
        owner_team:
          $ref: '#/components/schemas/VPCTeam'
        teams:
          type: array
          items:
            $ref: '#/components/schemas/VPCTeam'
        attached_vms_count:
          type: integer
          example: 5
      required:
        - id
        - name
        - description
        - date_created
        - status
        - location
        - v4_subnet
        - v4_subnet_mask
        - owner_team
        - teams
    VPCs:
      type: object
      properties:
        vpcs:
          type: array
          items:
            $ref: '#/components/schemas/VPC'
    AttachedVM:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: a1b2c3d4-5678-9abc-def0-123456789abc
        name:
          type: string
          example: "web-server-01"
        power_status:
          type: string
          example: "running"
        owner_team:
          $ref: '#/components/schemas/VPCTeam'
        mac:
          type: string
          example: "52:54:00:12:34:56"
        vpc_id:
          type: string
          format: uuid
          example: f5a3e2d8-9876-4abc-1234-543210fedcba
        vpc_name:
          type: string
          example: "Production VPC"
        address:
          type: string
          format: ipv4
          example: "10.0.0.10"
    VPCDetail:
      type: object
      properties:
        vpc:
          allOf:
            - $ref: '#/components/schemas/VPC'
            - type: object
              properties:
                attached_vms:
                  type: array
                  items:
                    $ref: '#/components/schemas/AttachedVM'
    NewVPC:
      type: object
      properties:
        team_id:
          type: string
          format: uuid
          example: c1d45377-b2a9-4407-ab8d-6909c34dfaac
          description: "Team ID. Default team will be used if not provided."
        location:
          type: string
          example: "sydney"
        name:
          type: string
          example: "Production VPC"
          maxLength: 200
        description:
          type: string
          example: "VPC for production environment"
          maxLength: 200
        v4_subnet:
          type: string
          example: "10.0.0.0"
          description: "IPv4 subnet prefix"
        v4_subnet_mask:
          type: string
          example: "24"
          description: "IPv4 subnet mask"
      required:
        - location
        - name
        - description
        - v4_subnet
        - v4_subnet_mask
    UpdateVPC:
      type: object
      properties:
        team_id:
          type: string
          format: uuid
          example: c1d45377-b2a9-4407-ab8d-6909c34dfaac
          description: "Team ID. Default team will be used if not provided."
        name:
          type: string
          example: "Updated VPC Name"
          maxLength: 255
        description:
          type: string
          example: "Updated VPC description"
          maxLength: 255
      required:
        - name
        - description
...
    FirewallRule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        group:
          type: string
          format: uuid
          description: Firewall group UUID
        ip_type:
          type: string
          enum: [v4, v6]
        action:
          type: string
          example: "allow"
        protocol:
          type: string
          enum: [tcp, udp, icmp, ipv6-icmp]
          description: "ICMP rules on a v6 subnet are stored and returned as ipv6-icmp"
        port:
          type: string
          example: "443"
          description: "Port number or range (e.g. '80', '8000:9000')"
        subnet:
          type: string
          example: "0.0.0.0"
          description: "IP address, CIDR, special value (Cloudflare, Onidel, etc.), or IP list reference (list:{uuid})"
        subnet_size:
          type: integer
          example: 0
          description: "CIDR prefix length"
        desc:
          type: string
          example: "Allow HTTPS"
    FirewallRules:
      type: object
      properties:
        firewall_rules:
          type: array
          items:
            $ref: '#/components/schemas/FirewallRule'
    FirewallRuleResponse:
      type: object
      properties:
        firewall_rule:
          $ref: '#/components/schemas/FirewallRule'
    NewFirewallRule:
      type: object
      properties:
        team_id:
          type: string
          format: uuid
        protocol:
          type: string
          enum: [tcp, udp, icmp, ipv6-icmp]
          example: "tcp"
          description: "icmp with a v6 subnet and ipv6-icmp are equivalent; the rule is stored with the version-appropriate name"
        port:
          type: string
          example: "443"
          description: "Port number or range (e.g. '80', '8000:9000'). Not required for ICMP."
        subnet:
          type: string
          example: "0.0.0.0"
          description: "IP/CIDR, special value (Cloudflare, Onidel, HetrixTools, CloudFront, UptimeRobot), or IP list (list:{uuid})"
        subnet_size:
          type: integer
          example: 0
          description: "CIDR prefix length (a numeric string is also accepted)"
        desc:
          type: string
          example: "Allow HTTPS traffic"
      required:
        - protocol
        - subnet
        - subnet_size
...
  securitySchemes:
    api_key:
      type: http
      scheme: bearer